### 2024-08-25 #7140 @dfm
@henrikmelin — Thanks for your submission! Given the small size of this code base, I have submitted this pre-review for editorial review with regards to the [substantial scholarly effort submission criterion](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). This usually takes a week or two and I'll report back here as soon as I hear back from the rest of the editorial board.
