### 2024-11-27 #7536 @editorialbot
[REVIEW]: `h3ppy`: An open-source Python package for modelling and fitting H$_3^+$ spectra
**Submitting author:** @henrikmelin (<a href="http://orcid.org/0000-0001-5971-2633">Henrik Melin</a>)
**Repository:** https://github.com/henrikmelin/h3ppy
**Branch with paper.md** (empty if default branch):
**Version:** 0.6.1
