### 2025-02-08 henrikmelin/h3ppy#1 @tedjohnson12
Correct import behavior
Per the JOSS review in https://github.com/openjournals/joss-reviews/issues/7536:

I ran this short script from the documentation:

``` python
import h3ppy
h3p = h3ppy.h3p()
wave = h3p.wavegen(3.4, 4.1, 1000)
h3p.set(T = 1000, N = 1e16, R = 2700, wave = wave)
m = h3p.model()
```

-   When installed from pip `pip install h3ppy==4.4.0` the script runs
    without error.
-   When installed from my local clone `v0.5.2`, there is no
    `h3ppy.h3p`. When the code became more than one file (sometime
    between `v0.4.4` and `v0.5.2`), an additional layer was added so
    that the first line of the above script must be changed to
    `from h3ppy import h3ppy`.

This can be fixed through the following action:

Give the current version of the code the correct import functionality
(i.e. `import h3ppy`). This can be done by removing line 37 in
`setup.py`:

``` python
package_dir={"h3ppy": ''},
```
