### 2025-02-08 henrikmelin/h3ppy#2 @tedjohnson12
Upload latest version to PyPi
Per the JOSS review in https://github.com/openjournals/joss-reviews/issues/7536:

The code is installable using pip, however, the latest version available
is `0.4.4`. The latest GitHub release is `0.5.2`.
