### 2025-02-08 henrikmelin/h3ppy#3 @tedjohnson12
Mistake in ``nirspec_jupiter.py``
https://github.com/openjournals/joss-reviews/issues/7536

See line 91 of ``examples/nirspec_jupiter.py``. Should `hp3.get_new_wavelength()` be `h3p.get_rest_wavelength()`? It ran normally after making this change.
