### 2025-02-08 henrikmelin/h3ppy#4 @tedjohnson12
Statement of need in documentation
https://github.com/openjournals/joss-reviews/issues/7536

Currently, there is a one-line description of the code. I would really like to see a description similar to the summary of the JOSS manuscript. That way, a potential user could understand the scope and goals of the code (and whether it is relevant to their work) before installation.
Per the review checklist, you should add a more detailed description to the documentation main page that clearly states what problems the code is designed to solve (including its scope) as well as the target audience.
