### 2025-02-08 henrikmelin/h3ppy#5 @tedjohnson12
[Recommendation] Add advanced examples to documentation page.
https://github.com/openjournals/joss-reviews/issues/7536

There is a basic example on the documentation page, but it is very minimal. The readme file contains many more detailed examples. These
examples would be much easier to find if they were on the documentation website. I recommend using the Sphinx plugin [Sphinx-gallery](https://sphinx-gallery.github.io/stable/index.html), but any other method would work.

This is mostly a recommendation because that is where I looked for examples, but I understand if you feel that they are better in the readme.
