### 2025-02-08 henrikmelin/h3ppy#6 @tedjohnson12
Clean up API reference
https://github.com/openjournals/joss-reviews/issues/7536

It is not clear what portions of the documented API are meant to be interacted with by users. Many of the functions are documented as accepting `**kwargs`, but it is not specified what keywords are acceptable. `kwargs` are fed into a function `h3p.parse_kwargs`, which edits the `vars` property of a `h3p` instance.

I recommend you cut down the number of class methods on the API reference page. In general, methods that are only called by in other methods should be marked \"private\" by prefixing the method name with `_`. For example, `h3p.parse_kwargs` should not be used by users, and so it should be changed to `h3p._parse_kwargs`. Sphinx automodule will recognize this as private and exclude it from the API reference page.

The portion of the API that is meant to be interacted with by users is mostly documented in a standard way. I think it would be useful to replace ``**kwargs`` in these function definitions with the actual allowed keywords. This will show up in the documentation and provide users with a way to quickly see what parameters they can include, and the keywords will be readable by a linter. (My own opinion is that you should also include type hints, but that is personal preference.)

For example,

``` python
def set(self, **kwargs) :
    """
    Use set() to set the model parameters that you require.

    .. code-block:: python

        h3p = h3ppy.h3p()
        h3p.set(temperature = 900, density = 1e15)

    Args:
            temperature: The temperature (in :math:`K`).
            density: The columnn densty (in :math:`m^{-2}`).
            wavelength: The wavelength scale (in :math:`\\mu m`).
            R: The spectral resolution (FWHM)

    .. hint::
        There are short-hand variable names for temperature (T), density (N), and wavelength (wave, w).

    """
    self.parse_kwargs(kwargs)
```

could be replaced by,

``` python
def set(
    self,
    temperature: float = None,
    density: float = None,
    wavelength: float = None,
    R: float = None,
    T: float = None,
    N: float = None,
    wave: float = None,
    R: float = None
):
    """
    Use set() to set the model parameters that you require.

    Args:
            temperature (float): The temperature (in :math:`K`).
            density (float): The columnn densty (in :math:`m^{-2}`).
            wavelength (float): The wavelength scale (in :math:`\\mu m`).
            R (float): The spectral resolution (FWHM)
            T (float): alias for temperature
            N (float): alias for density
            wave (float): alias for wavelength
            w (float): alias for wavelength
    """
    self.parse_kwargs({
        'temperature': temperature,
        'density': density,
        'wavelength': wavelength,
        'R': R,
        'T': T,
        'N': N,
        'wave': wave,
        'w': wave
    })
```
