### 2025-02-08 henrikmelin/h3ppy#7 @tedjohnson12
Comment on State of the Field in Software Paper
https://github.com/openjournals/joss-reviews/issues/7536

The paper does not compare `h3ppy` to other packages in the field. If there are none that fit H<sub>3</sub><sup>+</sup> spectra, then that should be stated.
