### 2025-02-08 #7536 @tedjohnson12
Hi all, sorry for the delay in completing the review. I've opened seven issues in the `h3ppy` repository with some actions that I think will make this code more accessible to new users. They include:

- https://github.com/henrikmelin/h3ppy/issues/1: Changing the import behavior match what is shown in the examples and in older versions of the code.
- https://github.com/henrikmelin/h3ppy/issues/2: Upload the latest release to PyPI
- https://github.com/henrikmelin/h3ppy/issues/3: Fix a typo in one example
- https://github.com/henrikmelin/h3ppy/issues/4: Add some more text to the main documentation page
- https://github.com/henrikmelin/h3ppy/issues/5: Adding examples to the documentation website
- https://github.com/henrikmelin/h3ppy/issues/6: Shaving down the API reference page to only include functions that the user will need to use
- https://github.com/henrikmelin/h3ppy/issues/7: Commenting on the state of the field in the software paper

With the exception of the first issue, none of these change the code functionality in any way. However, I hope they allow the user base to grow more quickly. Please let me know if you have questions!
