### 2025-02-14 #7536 @VincentKofman
Dear Dr. Henrik Melin,
Very nice work on the h3ppy software package. This is an extremely useful piece of work that is well documented and described. The documentation reads very well and I enjoyed working through the examples (you almost made it a little too easy!). I would like to congratulate you on a beautiful piece of work and have absolutely no hesitation to endorse the h3ppy package!
Best,
Vincent
