### 2025-03-03 bf3a84f
Resolves #3
Now calling the correct wavelength function.
