### 2025-03-03 henrikmelin/h3ppy#3 @henrikmelin
Thank you for catching this. Yes, it should be `h3p.get_rest_wavelength()`!
