### 2025-03-04 d6f7b0f
Resolves #4
JOSS intro and statement of need added.
