### 2025-03-04 henrikmelin/h3ppy#4 @henrikmelin
The overview and statement of need was added to the `index.rst` document page.
