### 2025-03-05 18952da
Added a logo for sphinx docs.
