+++ docs/conf.py
@@ -63,7 +63,7 @@ pygments_style = 'sphinx'
 html_theme = 'sphinx_rtd_theme'
 html_static_path = []
 # html_extra_path = ['google065a4d650d8ee82d.html']
-#html_logo = 'images/logo_rtd_transparent.png'
+html_logo = 'h3ppy-logo.png'
 html_theme_options = {
     'logo_only': True,
 }
