### 2025-03-05 2986c1d
Updates to docs, added examples, as per #5
