+++ docs/symbols.rst
@@ -2,28 +2,28 @@
 
 
 Definition of Symbols 
-*********************
+=====================
 
-=======================  ============================================  ==================================
-Symbol                   Parameter                                     Unit
-=======================  ============================================  ==================================
-:math:`I(\lambda, T)`    Spectral radiance		                       :math:`Wm^{-2} sr^{-1} \mu m^{-2}`
-:math:`Q(T)`             Partition function                            -- 
-:math:`R`                Spectral resolution                           -- 
-:math:`T`		         Temperature			                       :math:`K`
-:math:`N`                Column integrated :math:`H_3^+` density	   :math:`m^{-2}`
-:math:`b(\lambda)`		         Background level		                       :math:`Wm^{-2} sr^{-1} \mu m^{-2}`
-:math:`\lambda`	         Wavelength			                           :math:`\mu m`
-:math:`\sigma(\lambda)`  Wavelength dependent line width			   :math:`\mu m` 
-:math:`s(\lambda)`	     Wavelength dependent line shift			   :math:`\mu m` 
-:math:`g_{ns}`	         Spin weighting		                           --
-:math:`J`	             Angular momentum quantum nbr	               --
-:math:`w_{if}`	         Emission frequency                            :math:`cm^{-1}`
-:math:`w_{upper}` 	     Upper energy level		                       :math:`cm^{-1}`
-:math:`A_{if}`            Einstein A coefficient                       -- 
-:math:`h`                Planck's constant		                       :math:`m^{-2} kg s^{-1}`
-:math:`c`                Speed of light		                           :math:`m s^{-1}`
-:math:`k`                Boltzmann's constant                          :math:`m^2 kg s^{-2} K^{-1}`
-=======================  ============================================  ==================================
+=======================  =================================================  ==================================
+Symbol                   Parameter                                          Unit
+=======================  =================================================  ==================================
+:math:`I(\lambda, T)`    Spectral radiance		                            :math:`Wm^{-2} sr^{-1} \mu m^{-2}`
+:math:`Q(T)`             Partition function                                 -- 
+:math:`R`                Spectral resolution                                -- 
+:math:`T`		         Temperature			                            :math:`K`
+:math:`N`                Column integrated :math:`\text{H}_3^+` density	    :math:`m^{-2}`
+:math:`b(\lambda)`		 Background level		                            :math:`Wm^{-2} sr^{-1} \mu m^{-2}`
+:math:`\lambda`	         Wavelength			                                :math:`\mu m`
+:math:`\sigma(\lambda)`  Wavelength dependent line width			        :math:`\mu m` 
+:math:`s(\lambda)`	     Wavelength dependent line shift			        :math:`\mu m` 
+:math:`g_{ns}`	         Spin weighting		                                --
+:math:`J`	             Angular momentum quantum nbr	                    --
+:math:`w_{if}`	         Emission frequency                                 :math:`cm^{-1}`
+:math:`w_{upper}` 	     Upper energy level		                            :math:`cm^{-1}`
+:math:`A_{if}`            Einstein A coefficient                            -- 
+:math:`h`                Planck's constant		                            :math:`m^{-2} kg s^{-1}`
+:math:`c`                Speed of light		                                :math:`m s^{-1}`
+:math:`k`                Boltzmann's constant                               :math:`m^2 kg s^{-2} K^{-1}`
+=======================  =================================================  ==================================
 
 
