+++ docs/spectral_function.rst
@@ -8,7 +8,7 @@ Spectral Function
 
 
 
-The spectral radiance of :math:`H_3^+` at a particurlar wavelength can be calculated as a sum of guaussians, each representing an indivudial emission line.  
+The spectral radiance of :math:`\text{H}_3^+` at a particurlar wavelength can be calculated as a sum of guaussians, each representing an indivudial emission line.  
 
 .. math::
 
@@ -20,7 +20,7 @@ where :math:`I_i(T)` is the radiance of emission line :math:`i`. It is given by:
 
     I_i(T) = \frac{ g_{ns}(2J+1)100 \times hcw_{if}A_{if}}{4\pi Q(T)}\exp{\left[-\frac{100 \times hcw_{upper}}{kT}\right]}
 
-The :math:`H_3^+` partiation function :math:`Q(T)` is taken from Miller et al., (2013) and is expressed as a polynomial function: 
+The :math:`\text{H}_3^+` partiation function :math:`Q(T)` is taken from Miller et al., (2013) and is expressed as a polynomial function: 
 
 .. math::
 
@@ -28,4 +28,4 @@ The :math:`H_3^+` partiation function :math:`Q(T)` is taken from Miller et al., 
 
 where the constants :math:`a_n` are provided in Table 1 of Miller et al., (2013). 
 
-The :math:`H_3^+` line list is from Neale et al., (1996) provided in the package in a reduced form that removes emision lines with negligble radiance at temperatures < 2000 K. 
+The :math:`\text{H}_3^+` line list is from Neale et al., (1996) provided in the package in a reduced form that removes emision lines with negligble radiance at temperatures < 2000 K. 
