+++ docs/requirements.txt
@@ -1,3 +1,3 @@
-sphinx~=5.3
-sphinx_rtd_theme~=1.1
+sphinx==5.3.0
+sphinx_rtd_theme==1.1.1
 h3ppy
