+++ docs/installation.rst
@@ -7,11 +7,11 @@ Install `h3ppy` by using `pip`:
 
 .. code-block:: shell
 
-    pin install h3ppy
+    pip install h3ppy
 
 And update to the the latest version by
 
 .. code-block:: shell
 
-    pin install h3ppy --upgrade
+    pip install h3ppy --upgrade
 
