### 2025-03-07 db10594
Updated code as per #6
Changed some internal functions to be private so that they don't show in the documentation. Changed the behaviour of the set() function to have explicit keywords instead of just **kwargs. In light of these changes, the tests was also updated.
