+++ h3ppy/h3p.py
@@ -10,7 +10,7 @@ class h3p :
         The `set()`, `model()`, and `fit()` methods accepts the following inputs:
 
         * `wavelength`, `wave`, `w` - the wavelength scale on which to produce the model.  
-        * `data` - the observed H:math:`H_3^+` spectrum
+        * `data` - the observed :math:`H_3^+` spectrum
         * `temperature`, `T` - the intensity of the :math:`H_3^+` spectral lines are an exponential function of the temperature. Typical ranges for the ionosphere's of the giant planets are 400 (Saturn) to 1500 K (Jupiter). 
         * `density`, `N` - the column integrated :math:`H_3^+` density, this is the number of ions along the line of sight vector.
         * `sigma_n` - the nth polynomial constant of the spectral line width (sigma)
@@ -19,6 +19,7 @@ class h3p :
         * `nsigma` - the number of polynomial constant used for the sigma.
         * `noffset` - the number of polynomial constant used for the offset.
         * `nbackground` - the number of polynomial constant used for the background.
+        * `R` - the spectral resolving power of the instrument. This is converted in the code to a `sigma` parameter.
 
     """
     def __init__(self, line_list_file = '', **kwargs):
@@ -31,12 +32,12 @@ class h3p :
         else : self.line_list_file = line_list_file
 
         # Set up required parameters
-        self.startup()
+        self._startup()
 
         # Parse any potential input
-        self.parse_kwargs(kwargs)
+        self._parse_kwargs(kwargs)
 
-    def startup(self) : 
+    def _startup(self) : 
         """
             Perform a number of initialisations, define constants and default limits.
 
@@ -55,7 +56,7 @@ class h3p :
         self.h = 6.62607015e-34
 
         # Read the spectral information
-        self.read_line_list()
+        self._read_line_list()
 
         # This is the number of FWHM's before and after the line centre we'll
         # evaluate the intensity for. Purely for computational speedos.
@@ -84,7 +85,7 @@ class h3p :
         self._last_temperature = 0
         self._fit_sucess = False
 
-    def read_line_list(self) :
+    def _read_line_list(self) :
         """
             Read the line data from file into a structured array. 
 
@@ -102,7 +103,7 @@ class h3p :
         types = {'names' : ( 'Ju', 'wu', 'wl', 'EA', 'gw' ), 'formats' : ('f', 'f', 'f', 'f', 'f') }
         self.line_data = np.loadtxt(self.line_list_file, skiprows=1, dtype = types)
 
-    def poly_fn(self, pname, nbr) :
+    def _poly_fn(self, pname, nbr) :
         """
             Evaluate an polynomial function from the parameters in `self.var`.
     
@@ -129,7 +130,7 @@ class h3p :
                 The intensity of each individual line. 
         """
         
-        self.parse_kwargs(kwargs)
+        self._parse_kwargs(kwargs)
 
         # Don't recalculate if we've just done it
         if (self._last_temperature == self.vars['temperature']) :
@@ -161,16 +162,16 @@ class h3p :
                 A spectrum of units of spectral radiance (:math:`W m^{-2} sr^{-1} \\mu m^{-1}`). 
         """
 
-        self.parse_kwargs(kwargs)
+        self._parse_kwargs(kwargs)
 
         line_intensity = self.calculate_line_intensities()
 
-        self.background = self.poly_fn('background', self.nbackground)
+        self.background = self._poly_fn('background', self.nbackground)
 
-        return self.render_spectrum(line_intensity) * self.vars['density'] + self.background
+        return self._render_spectrum(line_intensity) * self.vars['density'] + self.background
 
     # Generate the spectrum
[…]
