+++ h3ppy/__init__.py
@@ -1,4 +1,4 @@
 from .h3p import h3p
 from .h2 import h2
 
-__version__ = 'v0.5.2'
+__version__ = 'v0.6.0'
