### 2025-03-07 85f8d8f
Commented on the state of the field. Resolves #7.
