+++ joss/paper.bib
@@ -1,7 +1,5 @@
 
 
-
-
 @article{Miller:2020,
 	adsnote = {Provided by the SAO/NASA Astrophysics Data System},
 	adsurl = {https://ui.adsabs.harvard.edu/abs/2020RvMP...92c5003M},
@@ -91,3 +89,40 @@
 	publisher = {McGraw-Hill},
 	title = {{Data reduction and error analysis for the physical sciences}},
 	year = 2003}
+
+@article{Dinelli:2017,
+	adsnote = {Provided by the SAO/NASA Astrophysics Data System},
+	adsurl = {https://ui.adsabs.harvard.edu/abs/2017GeoRL..44.4625D},
+	author = {{Dinelli}, B.~M. and {Fabiano}, F. and {Adriani}, A. and {Altieri}, F. and {Moriconi}, M.~L. and {Mura}, A. and {Sindoni}, G. and {Filacchione}, G. and {Tosi}, F. and {Migliorini}, A. and {Grassi}, D. and {Piccioni}, G. and {Noschese}, R. and {Cicchetti}, A. and {Bolton}, S.~J. and {Connerney}, J.~E.~P. and {Atreya}, S.~K. and {Bagenal}, F. and {Gladstone}, G.~R. and {Hansen}, C.~J. and {Kurth}, W.~S. and {Levin}, S.~M. and {Mauk}, B.~H. and {McComas}, D.~J. and {G{\`e}rard}, J. -C. and {Turrini}, D. and {Stefani}, S. and {Amoroso}, M. and {Olivieri}, A.},
+	date-added = {2025-01-10 15:58:36 +0000},
+	date-modified = {2025-01-10 15:58:36 +0000},
+	doi = {10.1002/2017GL072929},
+	journal = {\grl},
+	keywords = {Jupiter, aurora, Juno, JIRAM, H3+},
+	month = may,
+	number = {10},
+	pages = {4625-4632},
+	title = {{Preliminary JIRAM results from Juno polar observations: 1. Methodology and analysis applied to the Jovian northern polar region}},
+	volume = {44},
+	year = 2017,
+	bdsk-url-1 = {https://doi.org/10.1002/2017GL072929}}
+
+@article{Uno:2014,
+	adsnote = {Provided by the SAO/NASA Astrophysics Data System},
+	adsurl = {https://ui.adsabs.harvard.edu/abs/2014JGRA..11910219U},
+	author = {{Uno}, T. and {Kasaba}, Y. and {Tao}, C. and {Sakanoi}, T. and {Kagitani}, M. and {Fujisawa}, S. and {Kita}, H. and {Badman}, S.~V.},
+	date-added = {2024-07-04 11:17:20 +0100},
+	date-modified = {2024-07-04 11:17:20 +0100},
+	doi = {10.1002/2014JA020454},
+	journal = {Journal of Geophysical Research (Space Physics)},
+	keywords = {Jupiter, infrared aurora, H2, H3+, vertical profile},
+	month = dec,
+	number = {12},
+	pages = {10,219-10,241},
+	title = {{Vertical emissivity profiles of Jupiter's northern H$_{3}$$^{+}$ and H$_{2}$ infrared auroras observed by Subaru/IRCS}},
+	volume = {119},
+	year = 2014,
+	bdsk-url-1 = {https://doi.org/10.1002/2014JA020454}}
+
+
+
