+++ joss/paper.md
@@ -22,7 +22,7 @@ bibliography: paper.bib
 
 # Statement of need
 
-`h3ppy` seeks to simplify the process of analysing H$_3^+$ spectra by providing a standardised tool for the planetary science community. It is written in Python and installation is accessible via `pip`, which makes installing the package and maintainging it very straightforward. First-time users can generate a spectrum with only a few lines of code, whilst at the same time, the code provides more advanced control of the modelling and fitting process. 
+`h3ppy` seeks to simplify the process of analysing H$_3^+$ spectra by providing a standardised tool for the planetary science community. It is written in Python and installation is accessible via `pip`, which makes installing the package and maintaining it very straightforward. First-time users can generate a spectrum with only a few lines of code, whilst at the same time, the code provides more advanced control of the modelling and fitting process. Whilst other H$_3^+$ fitting codes exist in the literature [e.g., @Uno:2014, @Dinelli:2017], none are open-source. 
 
 For a given temperature and H$_3^+$ column density, the code calculates the radiance of each individual H$_3^+$ transition, then distributes these in wavelength-space by giving them Gaussian line shapes, the width of which is (principally) governed by the spectral resolving power of the spectrograph. The sum of all the individual Gaussian lines then make up the model spectrum. By making use of `numpy` array features (`numpy` is the only dependency), there are significant computational gains, making the code very fast. The fitting procedure uses a hard-coded least squares approach using partial derivatives of the spectral function, invoking Cramer's Rule [@Bevington:2003]. `h3ppy` implements the H$_3^+$ line list of @Neale:1996 and the partition function of @Miller:2010. 
 
@@ -32,6 +32,6 @@ The `h3ppy` code has already been used in published scientific studies, characte
 
 # Acknowledgements
 
-HM was supported by the STFC James Webb Fellowship (ST/W001527/2) at Northumbria University, UK.
+HM was supported by a STFC James Webb Fellowship (ST/W001527/2) at Northumbria University, UK.
 
 # References
