+++ joss/paper.bib
@@ -90,22 +90,6 @@
 	title = {{Data reduction and error analysis for the physical sciences}},
 	year = 2003}
 
-@article{Dinelli:2017,
-	adsnote = {Provided by the SAO/NASA Astrophysics Data System},
-	adsurl = {https://ui.adsabs.harvard.edu/abs/2017GeoRL..44.4625D},
-	author = {{Dinelli}, B.~M. and {Fabiano}, F. and {Adriani}, A. and {Altieri}, F. and {Moriconi}, M.~L. and {Mura}, A. and {Sindoni}, G. and {Filacchione}, G. and {Tosi}, F. and {Migliorini}, A. and {Grassi}, D. and {Piccioni}, G. and {Noschese}, R. and {Cicchetti}, A. and {Bolton}, S.~J. and {Connerney}, J.~E.~P. and {Atreya}, S.~K. and {Bagenal}, F. and {Gladstone}, G.~R. and {Hansen}, C.~J. and {Kurth}, W.~S. and {Levin}, S.~M. and {Mauk}, B.~H. and {McComas}, D.~J. and {G{\`e}rard}, J. -C. and {Turrini}, D. and {Stefani}, S. and {Amoroso}, M. and {Olivieri}, A.},
-	date-added = {2025-01-10 15:58:36 +0000},
-	date-modified = {2025-01-10 15:58:36 +0000},
-	doi = {10.1002/2017GL072929},
-	journal = {\grl},
-	keywords = {Jupiter, aurora, Juno, JIRAM, H3+},
-	month = may,
-	number = {10},
-	pages = {4625-4632},
-	title = {{Preliminary JIRAM results from Juno polar observations: 1. Methodology and analysis applied to the Jovian northern polar region}},
-	volume = {44},
-	year = 2017,
-	bdsk-url-1 = {https://doi.org/10.1002/2017GL072929}}
 
 @article{Uno:2014,
 	adsnote = {Provided by the SAO/NASA Astrophysics Data System},
@@ -124,5 +108,21 @@
 	year = 2014,
 	bdsk-url-1 = {https://doi.org/10.1002/2014JA020454}}
 
+@article{Dinelli:2017,
+	adsnote = {Provided by the SAO/NASA Astrophysics Data System},
+	adsurl = {https://ui.adsabs.harvard.edu/abs/2017GeoRL..44.4625D},
+	author = {{Dinelli}, B.~M. and {Fabiano}, F. and {Adriani}, A. and {Altieri}, F. and {Moriconi}, M.~L. and {Mura}, A. and {Sindoni}, G. and {Filacchione}, G. and {Tosi}, F. and {Migliorini}, A. and {Grassi}, D. and {Piccioni}, G. and {Noschese}, R. and {Cicchetti}, A. and {Bolton}, S.~J. and {Connerney}, J.~E.~P. and {Atreya}, S.~K. and {Bagenal}, F. and {Gladstone}, G.~R. and {Hansen}, C.~J. and {Kurth}, W.~S. and {Levin}, S.~M. and {Mauk}, B.~H. and {McComas}, D.~J. and {G{\`e}rard}, J. -C. and {Turrini}, D. and {Stefani}, S. and {Amoroso}, M. and {Olivieri}, A.},
+	date-added = {2025-01-10 15:58:36 +0000},
+	date-modified = {2025-01-10 15:58:36 +0000},
+	doi = {10.1002/2017GL072929},
+	journal = {\grl},
+	keywords = {Jupiter, aurora, Juno, JIRAM, H3+},
+	month = may,
+	number = {10},
+	pages = {4625-4632},
[…]
