+++ .readthedocs.yaml
@@ -15,8 +15,8 @@ sphinx:
   configuration: docs/conf.py
 
 # Optionally build your docs in additional formats such as PDF and ePub
-formats:
-  - pdf
+#formats:
+#  - pdf
 
 # Optional but recommended, declare the Python requirements required
 # to build your documentation
