### 2025-03-07 cbb263f
Updated version number to 0.6.0
