### 2025-03-07 henrikmelin/h3ppy#2 @henrikmelin
The new version (0.6.0) has been now published on pip.
