### 2025-03-07 henrikmelin/h3ppy#6 @henrikmelin
I've made some internal methods private, and have restructured the `set` function.
