### 2025-03-07 henrikmelin/h3ppy#5 @henrikmelin
The examples are now available on [readthedocs](https://h3ppy.readthedocs.io/en/latest/examples.html).
