### 2025-03-07 #7536 @henrikmelin
Hi @tedjohnson12 @VincentKofman! Many thanks for these comments!

I've worked on the Issues raised and have [closed them](https://github.com/henrikmelin/h3ppy/issues?q=is%3Aissue%20state%3Aclosed). I hope that I have successfully addressed the comments - please let me know if further changes are needed, happy to iterate.

The version of `h3ppy` that has the current changes from the JOSS reviews is now `0.6.0`, and is available on `pip`.

The updated documentation is [here](https://h3ppy.readthedocs.io/en/latest/index.html).

Thanks! 😃
