### 2025-03-08 #7536 @tedjohnson12
Hi @henrikmelin!  Thank you for addressing all of my comments. I agree with Vincent that this package will be useful to many people. Very happy to endorse it for publication!
Best,
Ted
