### 2025-03-09 #7536 @dfm
@henrikmelin — Thanks for your work responding to the feedback! Before final processing, can you update the manuscript to expand the `\grl` macro in the `.bib` file?

Once you've done that:

1. Take one last read through the manuscript to make sure that you're happy with it (it's harder to make changes later!), especially your name and affiliation. I've taken a pass and it looks good to me!
