+++ joss/paper.bib
@@ -115,7 +115,7 @@
 	date-added = {2025-01-10 15:58:36 +0000},
 	date-modified = {2025-01-10 15:58:36 +0000},
 	doi = {10.1002/2017GL072929},
-	journal = {\grl},
+	journal = {Geophysical Research Letters},
 	keywords = {Jupiter, aurora, Juno, JIRAM, H3+},
 	month = may,
 	number = {10},
