### 2025-03-10 76f3a8d
Updated version to 0.6.1
Version to archive on Zenodo for the JOSS paper.
