### 2025-03-10 #7536 @henrikmelin
Hi @dfm - brilliant thanks for this!

1. I've read through the manuscript, and I am happy :-)
2. The version number is `0.6.1`.
3. I've created an archived release of `0.6.1`, which can be accessed on [Zenodo](https://zenodo.org/records/14999082). I hope I've done this right - let me know if not!

Thanks!
