### 2025-03-12 #7536 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
