### 2025-03-13 #7536 @warrickball
Hi @henrikmelin, I'm the AEiC on this track. You can still change the paper source and those changes will be reflected in the published version. I'll start my final checks in the next few days and probably propose some editorial tweaks via PR anyway.

If you do make changes, you can still run `@editorialbot generate pdf` to have them rendered. You *won't* be able to run the `@editorialbot recommend-accept` command that does a test run of the publication pipeline.
