### 2025-03-15 aca6d38
Initial editorial tweaks for JOSS paper and documentation typos
