### 2025-03-15 #7536 @warrickball
Hi again @henrikmelin! I gather you already fixed your name in the paper. I've just opened [a pull request](https://github.com/henrikmelin/h3ppy/pull/8) with some editorial tweaks. If these are okay, let me know when you've merged them and I'll give the publication pipeline another test run.
