### 2025-03-16 #7536 @warrickball
Thanks! We're nearly there. I need to check if we can sort out two upper page limits for references being "+" and how exactly to handle $H_3^+$ in the title. Though it renders fine in the PDF, it leads to strange-looking metadata. I'll find out what the correct thing to do is and get back to you in the next few days. Sorry for the holdup!
