### 2025-03-27 9229cc9
Further editorial tweaks for JOSS paper
