+++ joss/paper.md
@@ -2,12 +2,12 @@
 title: 'h3ppy: An open-source Python package for modelling and fitting H$_3^+$ spectra'
 tags:
   - Python
-  - Astronomy
-  - Giant Planets
+  - astronomy
+  - giant planets
 authors:
   - name: Henrik Melin
     orcid: 0000-0001-5971-2633
-    affiliation: "1" 
+    affiliation: "1"
 affiliations:
  - name: Department of Mathematics, Physics, and Electrical Engineering, Northumbria University, Newcastle upon Tyne, UK
    index: 1
@@ -18,17 +18,17 @@ bibliography: paper.bib
 
 # Summary
 
-`h3ppy` is an open-source Python package for modelling and fitting the near-infrared spectrum of the tri-hydrogen cation H$_3^+$. This molecular ion is a main component of the charged particle ionospheres of the giant planets (Jupiter, Saturn, Uranus, and Neptune), and observations of these systems span over 30 years [@Miller:2020], using facilities such as ground-based telescopes (e.g. Keck, Very Large Telescope, NASA Infrared Telescope Facility), orbital spacecraft (e.g. Cassini and Juno), and space-based observatories (e.g., James Webb Space Telescope). By fitting the H$_3^+$ spectra, physical properties can be determined: 1) the temperature of the upper atmosphere and 2) the column integrated density of H$_3^+$ ions. The spatial and temporal distribution of these parameters reveal the processes and dynamics that govern the upper atmospheres of the giant planets, and in particular, how this region couples to both the lower atmosphere below and to the magnetic field beyond. `h3ppy` provides the tools required to both model the H$_3^+$ spectrum and perform these spectral retrievals. 
+`h3ppy` is an open-source Python package for modelling and fitting the near-infrared spectrum of the tri-hydrogen cation H$_3^+$. This molecular ion is a main component of the charged particle ionospheres of the giant planets (Jupiter, Saturn, Uranus, and Neptune), and observations of these systems span over 30 years [@Miller:2020], using facilities such as ground-based telescopes (e.g. Keck, Very Large Telescope, NASA Infrared Telescope Facility), orbital spacecraft (e.g. Cassini and Juno), and space-based observatories (e.g., James Webb Space Telescope). By fitting the H$_3^+$ spectra, physical properties can be determined: 1) the temperature of the upper atmosphere and 2) the column integrated density of H$_3^+$ ions. The spatial and temporal distribution of these parameters reveal the processes and dynamics that govern the upper atmospheres of the giant planets, and in particular, how this region couples to both the lower atmosphere below and to the magnetic field beyond. `h3ppy` provides the tools required to both model the H$_3^+$ spectrum and perform these spectral retrievals.
 
 # Statement of need
 
-`h3ppy` seeks to simplify the process of analysing H$_3^+$ spectra by providing a standardised tool for the planetary science community. It is written in Python and installation is accessible via `pip`, which makes installing the package and maintaining it very straightforward. First-time users can generate a spectrum with only a few lines of code, whilst at the same time, the code provides more advanced control of the modelling and fitting process. Whilst other H$_3^+$ fitting codes exist in the literature [e.g., @Dinelli:2017; @Uno:2014], none are open-source. 
+`h3ppy` seeks to simplify the process of analysing H$_3^+$ spectra by providing a standardised tool for the planetary science community. It is written in Python and installation is accessible via `pip`, which makes installing the package and maintaining it very straightforward. First-time users can generate a spectrum with only a few lines of code, whilst at the same time, the code provides more advanced control of the modelling and fitting process. Whilst other H$_3^+$ fitting codes exist in the literature [e.g., @Dinelli:2017; @Uno:2014], none are open-source.
 
[…]
