### 2025-03-27 #7536 @warrickball
Sorry for the delay. We've decided that there isn't any obviously better way to handle the title in the metadata. I've opened [another PR](https://github.com/henrikmelin/h3ppy/pull/9) with a few more tweaks I found while investigating the title but after that I think we'll be ready to publish.🤞
