### 2025-03-27 #7536 @henrikmelin
Thanks @warrickball! I've merged the PR.

What effect will the strange title metadata have? One (less preferred) solution is to rename the paper, substituting 'H$_3^+$' with 'tri-hydrogen cation'.
