### 2025-03-27 #7536 @warrickball
Congratulations @henrikmelin, your paper has been published in JOSS! 🎉
